requirements/test.txt
-r base.txt

flake8==3.5.0
pytest-cov==2.5.1
requests-mock==1.5.2
pytest-flask==0.10.0
